+++ sdk/core/core-amqp/src/errors.ts
@@ -7,6 +7,12 @@ import { isNode, isNumber, isString } from "../src/util/utils";
 import { isDefined, isObjectWithProperties } from "./util/typeGuards";
 
 /**
+ * The standard error message accompanying an AbortError.
+ * @hidden
+ */
+export const StandardAbortMessage = "The operation was aborted.";
+
+/**
  * Maps the conditions to the numeric AMQP Response status codes.
  * @internal
  */